Ajustes na validação do token. Ocultar o endoint interno.
