+++ api.py
@@ -1,10 +1,6 @@
-from services.validate import ValidateToken
 from routers import clients, favorites, token
-from dependencies import users_db, authenticate_user, create_access_token, ACCESS_TOKEN_EXPIRE_MINUTES, get_current_user
 
-from fastapi.security.oauth2 import OAuth2PasswordRequestForm
-from datetime import timedelta
-from fastapi import Depends, HTTPException, status, FastAPI
+from fastapi import FastAPI
 
 tags_metadata = [
     {
@@ -35,23 +31,4 @@ app.include_router(token.router)
 
 @app.get("/")
 async def read_main():
-#async def read_main(current_user: ValidateUser = Depends(get_current_user)):
     return {"msg": "Este projeto não contém front-end"}
-
-
-@app.post("/tokenauth", response_model=ValidateToken, tags=['tokenauth'])
-async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends()):
-    print(form_data)
-    user = authenticate_user(
-        users_db, form_data.username, form_data.password)
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Incorrect username or password",
-            headers={"WWW-Authenticate": "Bearer"},
-        )
-    access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
-    access_token = create_access_token(
-        data={"sub": user.username}, expires_delta=access_token_expires
-    )
-    return {"access_token": access_token, "token_type": "bearer"}
