+++ routers/token.py
@@ -1,7 +1,8 @@
-from services.validate import ValidateGenerateToken
+from services.validate import ValidateGenerateToken, ValidateToken
 from dependencies import users_db, authenticate_user, create_access_token, ACCESS_TOKEN_EXPIRE_MINUTES
 from datetime import timedelta
 
+from fastapi.security.oauth2 import OAuth2PasswordRequestForm
 from fastapi import APIRouter, Depends, HTTPException, status
 
 
@@ -15,6 +16,24 @@ router = APIRouter(prefix="/token",
 async def login_for_access_token(loc: ValidateGenerateToken = Depends()):
     user = authenticate_user(
         users_db, loc.username, loc.password)
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Incorrect username or password",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
+    access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
+    access_token = create_access_token(
+        data={"sub": user.username}, expires_delta=access_token_expires
+    )
+    return {"access_token": access_token, "token_type": "bearer"}
+
+
+@router.post("/tokenauth", response_model=ValidateToken, include_in_schema=False)
+async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends()):
+    print(form_data)
+    user = authenticate_user(
+        users_db, form_data.username, form_data.password)
     if not user:
         raise HTTPException(
             status_code=status.HTTP_401_UNAUTHORIZED,
